Moved functional tests for longest prefix matching into their own class and added JavaDoc, so as to keep the main functional test focused on the Car scenarios.
